+++ R/warp.R
@@ -81,7 +81,7 @@ transform_grid_grid = function(x, target) {
 	dxy = attr(target, "raster")$dimensions
 
 	from = st_crs(target)
-	pts = sf_project(from = from, to = st_crs(x), pts = new_pts)
+	pts = sf::sf_project(from = from, to = st_crs(x), pts = new_pts)
 
 	# at xy (target) locations, get values from x, or put NA
 	# to array:
@@ -159,7 +159,7 @@ st_warp = function(src, dest, ..., crs = NA_crs_, cellsize = NA_real_, segments 
 		options = c(options, "-dstnodata", no_data_value, "-r", method)
 		if (all(!is.na(cellsize))) {
 			cellsize = rep(abs(cellsize), length.out = 2)
-			options = c(options, "-ts", cellsize[1], cellsize[2])
+			options = c(options, "-tr", cellsize[1], cellsize[2])
 		}
 		if (! inherits(src, "stars_proxy")) {
 			src = st_as_stars_proxy(src, NA_value = no_data_value)
@@ -179,6 +179,9 @@ st_warp = function(src, dest, ..., crs = NA_crs_, cellsize = NA_real_, segments 
 			dest = tempfile(fileext = ".tif")
 			sf::gdal_utils("warp", src[[1]], dest, options = options)
 		} else {  # dest exists, and should be used: should use warper rather than warp
+			# https://github.com/r-spatial/stars/issues/407
+			if (length(dim(src)) == 3 && length(dim(dest)) == 2)
+				dest = merge(do.call(c, lapply(seq_len(dim(src)[3]), function(x) dest)))
 			dest = if (! inherits(dest, "stars_proxy")) {
 					dest[[1]] = NA_real_ * dest[[1]] # blank out values
 					delete = !debug